tests/wpt/tests/css/css-ui/historical/webkit-user-modify-01.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>-webkit-user-modify should be unsupported in author origin</title>
<link rel="help" href="https://developer.mozilla.org/en-US/docs/Web/CSS/user-modify">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#modified {
    -webkit-user-modify: read-write-plaintext-only;
}
</style>
<div id="inherited"></div>
<div id="modified"></div>
<script>
    test(function() {
        for (const el of [inherited, modified]) {
            assert_equals(el.style.WebkitUserModify, undefined);
            assert_equals(el.style.webkitUserModify, undefined);
            assert_equals(getComputedStyle(el).getPropertyValue('-webkit-user-modify'), '');
        }
    }, '-webkit-user-modify is not supported');
</script>
